Moved XeReplaceLLVMIntrinsics to separate file
